backend/requirements.txt
fastapi
uvicorn
groq
python-dotenv
SpeechRecognition
pydub
python-multipart
